requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-dotenv==1.0.0
alembic==1.11.1
redis==4.5.5
pydantic==1.10.9
transformers==4.30.2
torch==2.0.1
openai==0.27.8
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
spacy==3.5.3
nltk==3.8.1
pytest==7.3.1
celery==5.3.1
whisper==1.1.10
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
httpx==0.24.1
